+++ backend/SideQuest/src/main/java/com/main_032/SideQuest/community/Mapper/ChatMapper.java
@@ -1,0 +1,4 @@
+package com.main_032.SideQuest.community.Mapper;
+
+public class ChatMapper {
+}
