+++ backend/SideQuest/src/main/java/com/main_032/SideQuest/community/Mapper/CommentMapper.java
@@ -1,0 +1,4 @@
+package com.main_032.SideQuest.community.Mapper;
+
+public class CommentMapper {
+}
