+++ backend/SideQuest/src/main/java/com/main_032/SideQuest/community/dto/CommentDto/CommenntAriticle/CommentArticleRequest.java
@@ -1,31 +1,31 @@
-package com.main_032.SideQuest.community.entity.dto.CommentDto.CommentProject;
+package com.main_032.SideQuest.community.dto.CommentDto.CommenntAriticle;
 
-import com.main_032.SideQuest.community.entity.entity.Category;
+import com.main_032.SideQuest.community.entity.Category;
 import lombok.Getter;
 
 @Getter
-public class CommentProjectRequest {
+public class CommentArticleRequest {
     private Long memberId;
     private Category category;
     private String content;
     private int totalLikes;
-    private Long projectId;
+    private Long articleId; //assuming there's an articleId in Article entity
     private boolean deleted;
 
-    // CommentProjectRequest 생성자나 필요한 메소드가 있다면 여기에 추가하세요.
+    // CommentArticleRequest 생성자나 필요한 메소드가 있다면 여기에 추가하세요.
 
-    public CommentProjectRequest(
+    public CommentArticleRequest(
             Long memberId,
             Category category,
             String content,
             int totalLikes,
-            Long projectId,
+            Long articleId,
             boolean deleted) {
         this.memberId = memberId;
         this.category = category;
         this.content = content;
         this.totalLikes = totalLikes;
-        this.projectId = projectId;
+        this.articleId = articleId;
         this.deleted = deleted;
     }
 }
