+++ backend/SideQuest/src/main/java/com/main_032/SideQuest/community/dto/CommentDto/CommentProject/CommentProjectRequest.java
@@ -1,11 +1,10 @@
-package com.main_032.SideQuest.community.entity.dto.CommentDto.CommentProject;
+package com.main_032.SideQuest.community.dto.CommentDto.CommentProject;
 
-import com.main_032.SideQuest.community.entity.entity.Category;
+import com.main_032.SideQuest.community.entity.Category;
 import lombok.Getter;
 
 @Getter
-public class CommentProjectResponse {
-    private Long id;
+public class CommentProjectRequest {
     private Long memberId;
     private Category category;
     private String content;
@@ -13,17 +12,15 @@ public class CommentProjectResponse {
     private Long projectId;
     private boolean deleted;
 
-    // CommentProjectResponse 생성자나 필요한 메소드가 있다면 여기에 추가하세요.
+    // CommentProjectRequest 생성자나 필요한 메소드가 있다면 여기에 추가하세요.
 
-    public CommentProjectResponse(
-            Long id,
+    public CommentProjectRequest(
             Long memberId,
             Category category,
             String content,
             int totalLikes,
             Long projectId,
             boolean deleted) {
-        this.id = id;
         this.memberId = memberId;
         this.category = category;
         this.content = content;
