+++ backend/SideQuest/src/main/java/com/main_032/SideQuest/community/entity/Chat.java
@@ -1,4 +1,4 @@
-package com.main_032.SideQuest.community.entity.entity;
+package com.main_032.SideQuest.community.entity;
 
 import com.main_032.SideQuest.util.entity.BaseEntity;
 import lombok.Getter;
