--- backend/SideQuest/src/main/java/com/main_032/SideQuest/community/entity/Mapper/ChatMapper.java
@@ -1,4 +1,0 @@
-package com.main_032.SideQuest.community.entity.Mapper;
-
-public class ChatMapper {
-}
