--- backend/SideQuest/src/main/java/com/main_032/SideQuest/community/entity/dto/ChatDto.java
@@ -1,4 +1,0 @@
-package com.main_032.SideQuest.community.entity.dto;
-
-public class ChatDto {
-}
