--- backend/SideQuest/src/main/java/com/main_032/SideQuest/community/entity/repository/ChatRepository.java
@@ -1,4 +1,0 @@
-package com.main_032.SideQuest.community.entity.repository;
-
-public class ChatRepository {
-}
