--- backend/SideQuest/src/main/java/com/main_032/SideQuest/community/entity/service/ChatService.java
@@ -1,4 +1,0 @@
-package com.main_032.SideQuest.community.entity.service;
-
-public class ChatService {
-}
