+++ backend/SideQuest/src/main/java/com/main_032/SideQuest/community/repository/ChatRepository.java
@@ -1,0 +1,4 @@
+package com.main_032.SideQuest.community.repository;
+
+public class ChatRepository {
+}
