+++ backend/SideQuest/src/main/java/com/main_032/SideQuest/community/service/ChatService.java
@@ -1,0 +1,4 @@
+package com.main_032.SideQuest.community.service;
+
+public class ChatService {
+}
