+++ backend/SideQuest/src/main/java/com/main_032/SideQuest/community/service/CommentService.java
@@ -1,13 +1,11 @@
-package com.main_032.SideQuest.community.entity.service;
+package com.main_032.SideQuest.community.service;
 
 import com.main_032.SideQuest.article.entity.Article;
 import com.main_032.SideQuest.article.repository.ArticleRepository;
-import com.main_032.SideQuest.community.entity.dto.CommentDto.CommenntAriticle.CommentArticleDto;
-import com.main_032.SideQuest.community.entity.dto.CommentDto.CommentProject.CommentProjectDto;
-import com.main_032.SideQuest.community.entity.entity.Answer;
-import com.main_032.SideQuest.community.entity.entity.Comment;
-import com.main_032.SideQuest.community.entity.repository.AnswerRepository;
-import com.main_032.SideQuest.community.entity.repository.Commnet.CommentRepository;
+import com.main_032.SideQuest.community.dto.CommentDto.CommenntAriticle.CommentArticleDto;
+import com.main_032.SideQuest.community.dto.CommentDto.CommentProject.CommentProjectDto;
+import com.main_032.SideQuest.community.entity.Comment;
+import com.main_032.SideQuest.community.repository.Commnet.CommentRepository;
 import com.main_032.SideQuest.member.entity.Member;
 import com.main_032.SideQuest.member.repository.MemberRepository;
 import com.main_032.SideQuest.project.entity.Project;
